+++ tests/Mocks/SampleSubClass.php
@@ -7,4 +7,8 @@ class SampleSubClass extends SampleClass
     public function action(): void
     {
     }
+
+    public function action2(): void
+    {
+    }
 }